# tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "AgentOut": {
        "description": "Agent response for the chat UI.\n\nOnly exposes what the UI needs to render the agent selector and start a\nconversation. All internal fields (access_groups, enabled, timestamps)\nand infrastructure details (ARN, region, endpoint) are intentionally omitted.",
        "properties": {
          "agent_kind": {
            "title": "Agent Kind",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "multimodal": {
            "title": "Multimodal",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "suggested_questions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Questions"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "agent_kind",
          "multimodal",
          "suggested_questions"
        ],
        "title": "AgentOut",
        "type": "object"
      },
...
      "BreakdownOut": {
        "description": "Response body for POST /analytics/breakdown.\n\nArgs:\n    metric: The metric that was aggregated (cost, tokens, hits).\n    category: The category dimension used for ranking (user, agent, group).\n    labels: Ranked category labels in descending total order.\n    series: One or more named data series with values aligned to labels.",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/BreakdownSeriesItem"
            },
            "title": "Series",
            "type": "array"
          }
        },
        "required": [
          "metric",
          "category",
          "labels",
          "series"
        ],
        "title": "BreakdownOut",
        "type": "object"
      },
      "BreakdownRequest": {
        "description": "Request body for POST /analytics/breakdown.\n\nArgs:\n    from_date: Start date (inclusive), aliased as \"from\" in JSON.\n    to_date: End date (inclusive), aliased as \"to\" in JSON.\n    category: Dimension to rank by (user, agent, or group).\n    stack_by: Dimension to split series by (none, user, agent, group).\n    metric: Which metric to aggregate (cost, tokens, hits).\n    filters: Optional filter constraints restricting by user, agent, or group.\n    max_items: Maximum number of ranked category labels before folding into \"Other\".\n    max_series: Maximum number of distinct series before folding into \"Other\".",
        "properties": {
          "category": {
            "enum": [
              "user",
              "agent",
              "group"
            ],
            "title": "Category",
            "type": "string"
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FiltersIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "from": {
            "format": "date",
            "title": "From",
            "type": "string"
          },
          "max_items": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Items",
            "type": "integer"
          },
          "max_series": {
            "default": 10,
            "maximum": 25.0,
            "minimum": 1.0,
            "title": "Max Series",
            "type": "integer"
          },
          "metric": {
            "$ref": "#/components/schemas/Metric"
          },
          "stack_by": {
            "$ref": "#/components/schemas/Dimension"
          },
          "to": {
            "format": "date",
            "title": "To",
            "type": "string"
          }
        },
        "required": [
          "from",
          "to",
          "category",
          "stack_by",
          "metric"
        ],
        "title": "BreakdownRequest",
        "type": "object"
      },
      "BreakdownSeriesItem": {
        "description": "A single named data series in a breakdown response.\n\nArgs:\n    name: Display name of the series (metric name or stack_by dimension value).\n    values: Numeric values aligned positionally with the parent labels array.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "values": {
            "items": {
              "type": "number"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "name",
          "values"
        ],
        "title": "BreakdownSeriesItem",
        "type": "object"
      },
      "BudgetItemOut": {
        "description": "A single user's budget status for the current month.\n\nArgs:\n    user_id: User identifier.\n    username: Display name of the user.\n    spend: Total spend for the current UTC month.\n    budget: User's configured monthly budget.\n    pct_used: Fraction of budget consumed (0.0 when budget is 0).\n    status: Classification \u2014 \"ok\", \"warn\", or \"danger\".\n    is_over: Whether the user has exceeded their budget.",
        "properties": {
          "budget": {
            "title": "Budget",
            "type": "number"
          },
          "is_over": {
            "title": "Is Over",
            "type": "boolean"
          },
          "pct_used": {
            "title": "Pct Used",
            "type": "number"
          },
          "spend": {
            "title": "Spend",
            "type": "number"
          },
          "status": {
            "enum": [
              "ok",
              "warn",
              "danger"
            ],
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "spend",
          "budget",
          "pct_used",
          "status",
          "is_over"
        ],
        "title": "BudgetItemOut",
        "type": "object"
      },
      "BudgetStatusFilter": {
        "description": "Budget table filter options.",
        "enum": [
          "all",
          "approaching",
          "over"
        ],
        "title": "BudgetStatusFilter",
        "type": "string"
      },
      "Dimension": {
        "description": "Stack-by dimension for time-series and breakdown queries.",
        "enum": [
          "none",
          "user",
          "agent",
          "group"
        ],
        "title": "Dimension",
        "type": "string"
      },
      "FiltersIn": {
        "description": "Shared filter object for POST request bodies.\n\nArgs:\n    user_ids: Restrict results to sessions owned by these users.\n    agent_ids: Restrict results to sessions using these agents.\n    group_ids: Restrict results to sessions owned by users in these groups.",
        "properties": {
          "agent_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Ids"
          },
          "group_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Ids"
          },
          "user_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Ids"
          }
        },
        "title": "FiltersIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Interval": {
        "description": "Supported time bucket intervals for aggregation.",
        "enum": [
          "daily",
          "weekly",
          "monthly"
        ],
        "title": "Interval",
        "type": "string"
      },
...
      "MeUsageOut": {
        "description": "Response body for GET /users/me/usage.\n\nArgs:\n    interval: The interval used for bucketing (daily, weekly, monthly).\n    points: Ordered list of usage data points covering the full requested range.",
        "properties": {
          "interval": {
            "title": "Interval",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/MeUsagePoint"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "interval",
          "points"
        ],
        "title": "MeUsageOut",
        "type": "object"
      },
...
      "Metric": {
        "description": "Available metrics for timeline and breakdown queries.",
        "enum": [
          "cost",
          "tokens",
          "hits"
        ],
        "title": "Metric",
        "type": "string"
      },
...
      "SeriesItem": {
        "description": "A single named data series in a timeline response.\n\nArgs:\n    name: Display name of the series (metric name or dimension value).\n    data: Numeric values aligned positionally with the parent labels array.",
        "properties": {
          "data": {
            "items": {
              "type": "number"
            },
            "title": "Data",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "data"
        ],
        "title": "SeriesItem",
        "type": "object"
      },
...
      "SummaryOut": {
        "description": "Response body for POST /analytics/summary.\n\nArgs:\n    total_cost: Sum of TokenUsage.cost in the range.\n    input_tokens: Sum of TokenUsage.input_tokens in the range.\n    output_tokens: Sum of TokenUsage.output_tokens in the range.\n    hits: Count of assistant messages that are not errors in the range.\n    active_users: Count of distinct users with token usage in the range.",
        "properties": {
          "active_users": {
            "title": "Active Users",
            "type": "integer"
          },
          "hits": {
            "title": "Hits",
            "type": "integer"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "total_cost": {
            "title": "Total Cost",
            "type": "number"
          }
        },
        "required": [
          "total_cost",
          "input_tokens",
          "output_tokens",
          "hits",
          "active_users"
        ],
        "title": "SummaryOut",
        "type": "object"
      },
      "SummaryRequest": {
        "description": "Request body for POST /analytics/summary.\n\nArgs:\n    from_date: Start date (inclusive), aliased as \"from\" in JSON.\n    to_date: End date (inclusive), aliased as \"to\" in JSON.\n    filters: Optional filter constraints restricting by user, agent, or group.",
        "properties": {
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FiltersIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "from": {
            "format": "date",
            "title": "From",
            "type": "string"
          },
          "to": {
            "format": "date",
            "title": "To",
            "type": "string"
          }
        },
        "required": [
          "from",
          "to"
        ],
        "title": "SummaryRequest",
        "type": "object"
      },
      "TimelineOut": {
        "description": "Response body for POST /analytics/timeline.\n\nArgs:\n    interval: The interval used for bucketing (daily, weekly, monthly).\n    metric: The metric that was aggregated (cost, tokens, hits).\n    labels: Ordered time-bucket labels aligned with series data arrays.\n    series: One or more named data series.",
        "properties": {
          "interval": {
            "title": "Interval",
            "type": "string"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SeriesItem"
            },
            "title": "Series",
            "type": "array"
          }
        },
        "required": [
          "interval",
          "metric",
          "labels",
          "series"
        ],
        "title": "TimelineOut",
        "type": "object"
      },
      "TimelineRequest": {
        "description": "Request body for POST /analytics/timeline.\n\nArgs:\n    from_date: Start date (inclusive), aliased as \"from\" in JSON.\n    to_date: End date (inclusive), aliased as \"to\" in JSON.\n    interval: Time bucket granularity (daily, weekly, monthly).\n    metric: Which metric to aggregate (cost, tokens, hits).\n    stack_by: Dimension to split series by (none, user, agent, group).\n    filters: Optional filter constraints.\n    max_series: Maximum number of distinct series before folding into \"Other\".",
        "properties": {
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FiltersIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "from": {
            "format": "date",
            "title": "From",
            "type": "string"
          },
          "interval": {
            "$ref": "#/components/schemas/Interval"
          },
          "max_series": {
            "default": 10,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Series",
            "type": "integer"
          },
          "metric": {
            "$ref": "#/components/schemas/Metric"
          },
          "stack_by": {
            "$ref": "#/components/schemas/Dimension"
          },
          "to": {
            "format": "date",
            "title": "To",
            "type": "string"
          }
        },
        "required": [
          "from",
          "to",
          "interval",
          "metric",
          "stack_by"
        ],
        "title": "TimelineRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Strands Compose Chat",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agents": {
      "get": {
        "description": "List agents visible to the authenticated user.",
        "operationId": "list_agents_api_v1_agents_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentOut"
                  },
                  "title": "Response List Agents Api V1 Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Agents",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/{agent_id}": {
      "get": {
        "description": "Return a single agent visible to the authenticated user.",
        "operationId": "get_agent_api_v1_agents__agent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/analytics/breakdown": {
      "post": {
        "description": "Return ranked breakdown data for the admin analytics breakdown chart.\n\nArgs:\n    _admin: Authenticated admin user (used for auth gating only).\n    db: Async database session.\n    body: Breakdown request parameters including date range, category, metric,\n        and stack_by dimension.\n\nReturns:\n    BreakdownOut with ranked labels and one or more data series.",
        "operationId": "breakdown_handler_api_v1_analytics_breakdown_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreakdownRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreakdownOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Breakdown Handler",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/budgets": {
      "get": {
        "description": "Return per-user budget status for the current UTC month.\n\nArgs:\n    _admin: Authenticated admin user (used for auth gating only).\n    db: Async database session.\n    status: Filter by budget status \u2014 all, approaching, or over.\n    sort: Sort order \u2014 only \"pct_desc\" is accepted (422 otherwise).\n\nReturns:\n    List of BudgetItemOut sorted by pct_used descending.",
        "operationId": "budgets_handler_api_v1_analytics_budgets_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/BudgetStatusFilter",
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "const": "pct_desc",
              "default": "pct_desc",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BudgetItemOut"
                  },
                  "title": "Response Budgets Handler Api V1 Analytics Budgets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Budgets Handler",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/v1/analytics/summary": {
      "post": {
        "description": "Return aggregate KPI totals for the admin analytics summary cards.\n\nArgs:\n    _admin: Authenticated admin user (used for auth gating only).\n    db: Async database session.\n    body: Summary request parameters including date range and optional filters.\n\nReturns:\n    SummaryOut with total_cost, input_tokens, output_tokens, hits,\n    and active_users.",
        "operationId": "summary_handler_api_v1_analytics_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Summary Handler",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/timeline": {
      "post": {
        "description": "Return time-series data for the admin analytics timeline chart.\n\nArgs:\n    _admin: Authenticated admin user (used for auth gating only).\n    db: Async database session.\n    body: Timeline request parameters including date range, metric, and interval.\n\nReturns:\n    TimelineOut with labels and one or more data series.",
        "operationId": "timeline_handler_api_v1_analytics_timeline_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Timeline Handler",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/v1/users/me/usage": {
      "get": {
        "description": "Return pre-bucketed usage data for the authenticated user.\n\nArgs:\n    current_user: The authenticated user (injected, never from request params).\n    db: Async database session.\n    from_date: Start date (inclusive), parsed from \"from\" query param.\n    to_date: End date (inclusive), parsed from \"to\" query param.\n    interval: Time bucket granularity (daily, weekly, monthly).\n\nReturns:\n    MeUsageOut with interval and ordered list of usage data points.\n\nRaises:\n    ProblemDetailsException: HTTP 422 when from_date > to_date.",
        "operationId": "me_usage_handler_api_v1_users_me_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Interval"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeUsageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me Usage Handler",
        "tags": [
          "usage"
        ]
      }
    },
...
  }
}